Put ferc1_eia training functions in a new and different module called ferc1_eia_train instead of all in ferc1_eia.
Create a notebook specifically for FERC-EIA manual mapping, validation, and integration with current training data. Second notebook for creating the spreadsheets also created.

Remove training data from IDACORP, Southern, and Evergy because there were duplicates

Updated the path references in the ferc1_eia_train module. Still feels janky, but it works
